+++ choco-solver/src/main/java/org/chocosolver/solver/constraints/nary/nogood/NogoodConstraint.java
@@ -1,24 +1,22 @@
-/**
- * Copyright (c) 2014,
- *       Charles Prud'homme (TASC, INRIA Rennes, LINA CNRS UMR 6241),
- *       Jean-Guillaume Fages (COSLING S.A.S.).
+/*
+ * Copyright (c) 1999-2015, Ecole des Mines de Nantes
  * All rights reserved.
- *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
+ *
  *     * Redistributions of source code must retain the above copyright
  *       notice, this list of conditions and the following disclaimer.
  *     * Redistributions in binary form must reproduce the above copyright
  *       notice, this list of conditions and the following disclaimer in the
  *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the <organization> nor the
+ *     * Neither the name of the Ecole des Mines de Nantes nor the
  *       names of its contributors may be used to endorse or promote products
  *       derived from this software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * THIS SOFTWARE IS PROVIDED BY THE REGENTS AND CONTRIBUTORS ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+ * DISCLAIMED. IN NO EVENT SHALL THE REGENTS AND CONTRIBUTORS BE LIABLE FOR ANY
  * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
  * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
  * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
@@ -28,40 +26,30 @@
  */
 package org.chocosolver.solver.constraints.nary.nogood;
 
-import org.chocosolver.solver.exception.ContradictionException;
-import org.chocosolver.solver.variables.IntVar;
+import org.chocosolver.solver.Solver;
+import org.chocosolver.solver.constraints.Constraint;
+import org.chocosolver.solver.constraints.nary.cnf.PropNogoods;
 import org.chocosolver.util.ESat;
 
 /**
- * An interface to define a Nogood.
- * <p/>
- * Made of a list of variables, a list of values and an int.
- * {vars, values} matches positive decisions.
- * A positive decision d_i is vars_i=values_i.
- * <p/>
- * Related to "Nogood Recording from Restarts", C. Lecoutre et al.
- * <br/>
- *
- * @author Charles Prud'homme
- * @since 20/06/13
+ * Created by cprudhom on 21/01/15.
+ * Project: choco.
  */
-public interface INogood {
+public class NogoodConstraint extends Constraint {
 
-    void setIdx(int idx);
+    final PropNogoods nogoods;
 
-    int getIdx();
+    public NogoodConstraint(Solver solver) {
+        super("NogoodConstraint", new PropNogoods(solver));
+        nogoods = (PropNogoods) propagators[0];
+    }
 
-    int propagate(PropNogoodStore pngs) throws ContradictionException;
+    @Override
+    public ESat isSatisfied() {
+        return nogoods.isEntailed();
+    }
 
-    int awakeOnInst(int idx, PropNogoodStore pngs) throws ContradictionException;
-
-    boolean isUnit();
-
-    ESat isEntailed();
-
-    int size();
-
-    IntVar getVar(int i);
-
-    int getVal(int i);
+    public PropNogoods getPropNogoods() {
+        return nogoods;
+    }
 }