+++ choco-solver/src/main/java/org/chocosolver/solver/propagation/IPropagationEngine.java
@@ -77,25 +77,33 @@ public interface IPropagationEngine extends Serializable {
      *
      * @return true if the engine has been initialized
      */
-    boolean isInitialized();
+    default boolean isInitialized() {
+        return false;
+    }
 
     /**
      * Launch the proapagation, ie, active propagators if necessary, then reach a fix point
      *
      * @throws ContradictionException if a contradiction occurrs
      */
-    void propagate() throws ContradictionException;
+    default void propagate() throws ContradictionException {
+    }
 
     /**
      * Flush <code>this</code>, ie. remove every pending events
      */
-    void flush();
+    default void flush() {
+    }
 
-    void fails(ICause cause, Variable variable, String message) throws ContradictionException;
+    default void fails(ICause cause, Variable variable, String message) throws ContradictionException {
+    }
 
-    ContradictionException getContradictionException();
+    default ContradictionException getContradictionException() {
+        throw new UnsupportedOperationException("no propagation engine has been defined");
+    }
 
-    void clear();
+    default void clear() {
+    }
 
     //********************************//
     //      SERVICES FOR UPDATING     //
@@ -105,38 +113,58 @@ public interface IPropagationEngine extends Serializable {
      * Take into account the modification of a variable
      *
      * @param variable modified variable
-     * @param type type of modification event
+     * @param type     type of modification event
      * @throws ContradictionException
      */
-    void onVariableUpdate(Variable variable, IEventType type, ICause cause) throws ContradictionException;
+    default void onVariableUpdate(Variable variable, IEventType type, ICause cause) throws ContradictionException {
+    }
 
-    void delayedPropagation(Propagator propagator, PropagatorEventType type) throws ContradictionException;
+    default void delayedPropagation(Propagator propagator, PropagatorEventType type) throws ContradictionException {
+    }
 
-    void onPropagatorExecution(Propagator propagator);
+    default void onPropagatorExecution(Propagator propagator) {
+    }
 
     /**
      * Set the propagator as inactivated within the propagation engine
      *
      * @param propagator propagator to desactivate
      */
-    void desactivatePropagator(Propagator propagator);
+    default void desactivatePropagator(Propagator propagator) {
+    }
 
     /**
      * Add a constraint to the propagation engine
+     *
      * @param permanent does the constraint is permanently added
-     * @param ps propagators to add
+     * @param ps        propagators to add
      */
-    void dynamicAddition(boolean permanent, Propagator... ps);
+    default void dynamicAddition(boolean permanent, Propagator... ps) {
+    }
 
     /**
      * Update the scope of variable of a propagator (addition or deletion are allowed -- p.vars are scanned)
+     *
      * @param p a propagator
      */
-    void updateInvolvedVariables(Propagator p);
+    default void updateInvolvedVariables(Propagator p) {
+    }
+
+
+    /**
+     * State that the propagator needs to be propagated (coarse event) on backtrack.
+     * For dynamic propagator only, such as PropSat or PropNogoods.
+     *
+     * @param p a propagator
+     */
+    default void propagateOnBacktrack(Propagator p) {
+    }
 
     /**
      * Delete the list of propagators in input from the engine
+     *
      * @param ps a list of propagators
      */
-    void dynamicDeletion(Propagator... ps);
+    default void dynamicDeletion(Propagator... ps) {
+    }
 }